move functions.js to modules
